linux: Move `futex.h` API to `src/new`
(backport <https://github.com/rust-lang/libc/pull/5217>)
(cherry picked from commit a8e295810877b50961f1217330108a8c37505fe0)
