+++ src/new/linux_uapi/linux/mod.rs
@@ -3,6 +3,7 @@
 //! <https://github.com/torvalds/linux/tree/master/include/uapi/linux>
 
 pub(crate) mod can;
+pub(crate) mod futex;
 pub(crate) mod if_link;
 pub(crate) mod if_packet;
 pub(crate) mod keyctl;
