+++ src/unix/linux_like/mod.rs
@@ -250,14 +250,6 @@ cfg_if! {
                 pub len: c_ushort,
                 pub filter: *mut sock_filter,
             }
-
-            // linux/futex.h
-            pub struct futex_waitv {
-                pub val: crate::__u64,
-                pub uaddr: crate::__u64,
-                pub flags: crate::__u32,
-                __reserved: Padding<crate::__u32>,
-            }
         }
     }
 }
